move methods to DataManager
